refactor: test를 위해 prisma 경로  변경 #20
