+++ src/domain/member/member.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { MemberController } from './member.controller';
 import { MemberService } from './member.service';
 import { MemberRepository } from './member.repository';
-import { PrismaModule } from 'src/global/config/prisma/prisma.module';
+import { PrismaModule } from '../../global/database/prisma/prisma.module';
 
 @Module({
   imports: [PrismaModule],
