+++ src/domain/member/member.repository.ts
@@ -2,7 +2,7 @@ import { Injectable, Logger } from '@nestjs/common';
 import { LoginType, Role } from '@prisma/client';
 import { randomUUID } from 'crypto';
 import { IAuthSocial } from './interface/member.interface';
-import { PrismaService } from '../../global/config/prisma/prisma.service';
+import { PrismaService } from '../../global/database/prisma/prisma.service';
 
 @Injectable()
 export class MemberRepository {
